requirements.txt
iso3166==2.0.1
stripe==2.60.0
Django==3.2.6
djangorestframework==3.12.4
Pillow==8.3.2
